// src/types/generated/contentful.d.ts
export interface IAuthorFields {
  /** ID */
  id?: string | undefined;

  /** First Name */
  firstName?: string | undefined;

  /** Last Name */
  lastName?: string | undefined;

  /** Email */
  email?: string | undefined;

  /** Biography */
  biography?: Document | undefined;

  /** Profile Picture */
  profilePicture?: Asset | undefined;

  /** Job */
  job?: string | undefined;
}

/** firstname lastname email biography profilePicture job id */

export interface IAuthor extends Entry<IAuthorFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "author";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IBlogFields {
  /** Id */
  id: string;

  /** Title */
  title: string;

  /** Description */
  description?: string | undefined;

  /** Body */
  body?: Document | undefined;

  /** Slug */
  slug: string;

  /** Featured Image */
  featuredImage?: Asset | undefined;

  /** Category */
  category?:
    | "Travel"
    | "Food and Cooking"
    | "Health and Wellness"
    | "Technology"
    | "Lifestyle"
    | "Personal Development"
    | "Finance and Money"
    | "Parenting"
    | "Books and Literature"
    | "Travel Photography"
    | "Gaming"
    | "DIY and Crafts"
    | "Entrepreneurship"
    | "Music"
    | undefined;

  /** Tags */
  tags?: string[] | undefined;

  /** Author */
  author?: IAuthor | undefined;
}

export interface IBlog extends Entry<IBlogFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "blog";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}